refactor: GNB4 Match 폴더구조 정리
