+++ src/pages/match/MercenaryContent.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import MercenaryArticleCard from './MercenaryArticleCard';
+import MercenaryArticleCard from './mercenary_feature/MercenaryArticleCard';
 import styled from 'styled-components';
 
 const Container = styled.div`
