+++ src/pages/match/mercenary_feature/MercenaryHistory.jsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import styled from 'styled-components';
-import previous_icon from '../../assets/icons/arrow-left.svg';
+import previous_icon from '../../../assets/icons/arrow-left.svg';
 import MercenaryReceivedList from './MercenaryReceivedList';
 import MercenarySentList from './MercenarySentList'
